requirements.txt
dotenv==0.9.9
paho-mqtt==2.1.0
PyMySQL==1.1.2
python-dotenv==1.2.1
